Move tray_icon fixture
